Update design almost finish

+++ src/app/pages/list-vivencias/list-vivencias.component.ts
@@ -1,18 +1,26 @@
 import { Component, Input, OnInit } from '@angular/core';
+import { AlertController } from '@ionic/angular';
 import { Vivencia } from 'src/app/interfaces/vivencia';
-import { CrudStorageService } from 'src/app/services/crud-storage.service';
+import { CrudService } from 'src/app/services/crud-storage.service';
 
 @Component({
   selector: 'app-list-vivencias',
   templateUrl: './list-vivencias.component.html',
   styleUrls: ['./list-vivencias.component.scss'],
-  providers: [CrudStorageService]
+  providers: [CrudService]
 })
 export class ListVivenciasComponent  implements OnInit {
 
   @Input() listOfVivencias!: Vivencia[];
 
-  constructor( private _crudStorage: CrudStorageService ) {
+  handlerMessage: string = '';
+  roleMessage: string = '';
+  confirm: string | undefined;
+
+  constructor(
+    private _crudStorage: CrudService,
+    private alertController: AlertController
+    ) {
   }
 
   ngOnInit() {
@@ -29,9 +37,45 @@ export class ListVivenciasComponent  implements OnInit {
   }
 
   clearData() {
-    this._crudStorage.delete().then(_data => {
-      this.listOfVivencias = [];
-    })
+    this.presentAlert().then(rs => {
+      if(this.confirm !== "confirm"){
+        console.log("confirm", this.confirm);
+        return;
+      }
+      this._crudStorage.delete().then(_data => {
+        this.listOfVivencias = [];
+      });
+    });
+
+  }
+
+  async presentAlert() {
+    const alert = await this.alertController.create({
+      header: 'Confirmar',
+      message: "Seguro que desea eliminar las vivencias?",
+      buttons: [
+        {
+          text: 'Cancel',
+          role: 'cancel',
+          handler: () => {
+            this.handlerMessage = 'Alert canceled';
+          },
+        },
+        {
+          text: 'Yes',
+          role: 'confirm',
+          handler: () => {
+            this.handlerMessage = 'Alert confirmed';
+          },
+        },
+      ],
+    });
+
+    await alert.present();
+
+    const { role } = await alert.onDidDismiss();
+    this.confirm = role;
+    this.roleMessage = `Dismissed with role: ${role}`;
   }
 
 }
